Move registry to cli
